Package bayessource

Bayesian approach to evaluate whether two sets of multivariate observations come from the same source.

Details

The observations are assumed to be generated by a hierarchical Normal-Inverted Wishart distribution. The hyperparameters can be fitted using additional background data, covering samples from multiple sources.

The package implements a Gibbs sampler to sample from the posteriors, and the computation of the marginal likelihood follows Chib (1995). The Bayes factor can also be computed as a ratio of two marginal likelihoods.

Normal-Inverse-Wishart model

Described in (Bozza et al. 2008) .

Observation level:

$$X_{ij} \sim N_p(\theta_i, W_i)$$ (i = source, j = items from source)

Group level:

$$\theta_i \sim N_p(\mu, B)$$

$$W_i \sim IW_p(\nu_w, U)$$

Hyperparameters:

$$B, U, \nu_w, \mu$$

Posterior samples of \(\theta\), \(W^{(-1)}\) can be generated with a Gibbs sampler.

References

Bozza S, Taroni F, Marquis R, Schmittbuhl M (2008). “Probabilistic Evaluation of Handwriting Evidence: Likelihood Ratio for Authorship.” Journal of the Royal Statistical Society: Series C (Applied Statistics), 57(3), 329-341. ISSN 1467-9876, doi: 10.1111/j.1467-9876.2007.00616.x .

See also