Wishart density.

Density of X ~ Wishart(df, Sigma).

dwishart(X, df, Sigma, log = FALSE, is.chol = FALSE)

Arguments

X
the observation

df
degrees of freedom

Sigma
scale matrix

log
if TRUE, return the log-density

is.chol
if TRUE, Sigma and X are the upper Cholesky factors of Sigma and X

Value

the density in X

Wishart parametrization

Uses (Press 2012; Anderson 2003) parametrization.

$$X ~ W(v, S)$$ with \(S = pxp\) matrix, \(v >= p\) (the degrees of freedom).

Then: $$E[X] = v * S$$

References

Anderson TW (2003). An Introduction to Multivariate Statistical Analysis, Wiley Series in Probability and Statistics, 3rd ed edition. Wiley-Interscience, Hoboken, N.J. ISBN 978-0-471-36091-9.

Press SJ (2012). Applied Multivariate Analysis: Using Bayesian and Frequentist Methods of Inference. Courier Corporation.