Elicit priors and initialization from background dataset

Here we fix the hyperparameters for priors on \(theta_i\) and \(W_i\), i.e., \(B\), \(U\), \(mu\) and \(n_w\).

make_priors_and_init(df.background, col.variables, col.item,
  use.priors = "ML", use.init = "random", ...)

Arguments

df.background
the background dataset

col.variables
columns with variables

col.item
column with item id

use.priors
see details

use.init
see details

...
additional variables for priors, init

Value

a list of variables

Details

Here we fix the hyperparameters for priors on \(theta_i\) and \(W_i\), i.e., \(B\), \(U\), \(mu\) and \(n_w\).

An appropriate initialization value for \(W^{-1}_i\), \(i=1,2\) is also generated.

Notice that we have three chains in the LR computations, the same initialization is used thrice.

Priors

use.priors:

'ML': maximum likelihood estimation

'vague': low-information priors

U is alpha*diag(p), B is beta*diag(p), mu is mu0. By default alpha = 1, beta = 1, mu0 = 0

The Wishart dofs nw are set as small as possible without losing full rank: \(nw = 2*(p + 1) + 1\)

Initialization

use.init:

'random': initialize according to the model

'vague': low-information initialization \(W_i\) is alpha_init + beta_init *diag(p)

Some constants can be changed by passing the new values to ... :

use.priors = 'vague': alpha = 1, beta = 1, mu0 = 0

use.init = 'vague': alpha_init = 1, beta_init = 100

Returns

A list of variables:

mu: the global mean

B.inv: the between prior covariance, as the inverse

U: the Inverted Wishart scale matrix

nw: the Inverted Wishart dof

W.inv.1, W.inv.2: the within cov initializations, as inverses